refactor: Separate the server.js file
