+++ src/server/init.js
@@ -1,0 +1,18 @@
+import jsonServer from 'json-server';
+import { Low, JSONFile } from 'lowdb';
+import addAction from './middleware.js';
+
+const server = jsonServer.create();
+const middlewares = jsonServer.defaults();
+const router = jsonServer.router('db.json');
+const adapter = new JSONFile('db.json');
+const db = new Low(adapter);
+
+await db.read();
+
+server.use(middlewares);
+server.use(jsonServer.bodyParser);
+server.use(router);
+server.use(addAction(db));
+
+export default server;
